Stop using ServerStartOptions in other commands
