+++ core/Azure.Mcp.Core/tests/Azure.Mcp.Core.Tests/Areas/Server/Commands/ToolLoading/ServerToolLoaderTests.cs
@@ -56,10 +56,10 @@ public class ServerToolLoaderTests
         // Arrange - use real RegistryDiscoveryStrategy since ServerToolLoader depends on it
         var serviceProvider = new ServiceCollection().AddLogging().BuildServiceProvider();
         var loggerFactory = serviceProvider.GetRequiredService<ILoggerFactory>();
-        var serviceStartOptions = Microsoft.Extensions.Options.Options.Create(new ServerStartOptions());
+        var serverStartOptions = Microsoft.Extensions.Options.Options.Create(new ServerStartOptions());
         var toolLoaderOptions = Microsoft.Extensions.Options.Options.Create(new ToolLoaderOptions());
         var discoveryLogger = loggerFactory.CreateLogger<RegistryDiscoveryStrategy>();
-        var discoveryStrategy = RegistryDiscoveryStrategyHelper.CreateStrategy(serviceStartOptions.Value, discoveryLogger);
+        var discoveryStrategy = RegistryDiscoveryStrategyHelper.CreateStrategy(serverStartOptions.Value, discoveryLogger);
         var logger = loggerFactory.CreateLogger<ServerToolLoader>();
 
         var toolLoader = new ServerToolLoader(discoveryStrategy, toolLoaderOptions, logger);
@@ -110,10 +110,10 @@ public class ServerToolLoaderTests
         // Arrange - use real RegistryDiscoveryStrategy
         var serviceProvider = new ServiceCollection().AddLogging().BuildServiceProvider();
         var loggerFactory = serviceProvider.GetRequiredService<ILoggerFactory>();
-        var serviceStartOptions = Microsoft.Extensions.Options.Options.Create(new ServerStartOptions());
+        var serverStartOptions = Microsoft.Extensions.Options.Options.Create(new ServerStartOptions());
         var toolLoaderOptions = Microsoft.Extensions.Options.Options.Create(new ToolLoaderOptions());
         var discoveryLogger = loggerFactory.CreateLogger<RegistryDiscoveryStrategy>();
-        var discoveryStrategy = RegistryDiscoveryStrategyHelper.CreateStrategy(serviceStartOptions.Value, discoveryLogger);
+        var discoveryStrategy = RegistryDiscoveryStrategyHelper.CreateStrategy(serverStartOptions.Value, discoveryLogger);
         var logger = loggerFactory.CreateLogger<ServerToolLoader>();
 
         var toolLoader = new ServerToolLoader(discoveryStrategy, toolLoaderOptions, logger);
