+++ servers/Fabric.Mcp.Server/src/Program.cs
@@ -189,7 +189,7 @@ internal class Program
         // !!! WARNING !!!
         // stdio-transport-specific implementations of ICacheService.
         // The http-transport-specific implementations and configurations must be registered
-        // within ServiceStartCommand.ExecuteAsync().
+        // within ServerStartCommand.ExecuteAsync().
         services.AddHttpClientServices();
         services.AddSingleUserCliCacheService(disabled: true);
