+++ tools/Azure.Mcp.Tools.Compute/tests/Azure.Mcp.Tools.Compute.Tests/Vm/VmCreateCommandTests.cs
@@ -61,6 +61,7 @@ public class VmCreateCommandTests : SubscriptionCommandUnitTestsBase<VmCreateCom
                 Tags: null);
 
             Service.CreateVmAsync(
+                Arg.Any<bool>(),
                 Arg.Any<string>(),
                 Arg.Any<string>(),
                 Arg.Any<string>(),
@@ -119,6 +120,7 @@ public class VmCreateCommandTests : SubscriptionCommandUnitTestsBase<VmCreateCom
             Tags: new Dictionary<string, string> { { "env", "test" } });
 
         Service.CreateVmAsync(
+            Arg.Any<bool>(),
             Arg.Is(_knownVmName),
             Arg.Is(_knownResourceGroup),
             Arg.Is(_knownSubscription),
@@ -185,6 +187,7 @@ public class VmCreateCommandTests : SubscriptionCommandUnitTestsBase<VmCreateCom
         var conflictException = new RequestFailedException((int)HttpStatusCode.Conflict, "A VM with this name already exists");
 
         Service.CreateVmAsync(
+            Arg.Any<bool>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
@@ -231,6 +234,7 @@ public class VmCreateCommandTests : SubscriptionCommandUnitTestsBase<VmCreateCom
         var forbiddenException = new RequestFailedException((int)HttpStatusCode.Forbidden, "Authorization failed");
 
         Service.CreateVmAsync(
+            Arg.Any<bool>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
@@ -287,6 +291,7 @@ public class VmCreateCommandTests : SubscriptionCommandUnitTestsBase<VmCreateCom
             Tags: null);
 
         Service.CreateVmAsync(
+            Arg.Any<bool>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
             Arg.Any<string>(),
